Refactor optimizers and implement Nova optimizer
- Moved each optimizer to its own file under `pure_numpy_nn/optimizers`.
- Implemented a new optimizer called `Nova`.
- Nova combines momentum, RMSProp, and a sign-consistency trust factor.
- Benchmarked Nova against Adam and found it to be slightly better on the given task.
